Remove empty borough and error in integer division in assignment 3

+++ src/main/java/accidentsBorough/AccidentsBoroughReducer.java
@@ -9,19 +9,21 @@ import org.apache.hadoop.mapred.OutputCollector;
 import org.apache.hadoop.mapred.Reducer;
 import org.apache.hadoop.mapred.Reporter;
 
-public class AccidentsBoroughReducer extends MapReduceBase implements Reducer<WeekBoroughWritable, IntWritable, WeekBoroughWritable, AvgAccidentsAndLetalWritable> {
+public class AccidentsBoroughReducer extends MapReduceBase
+		implements Reducer<WeekBoroughWritable, IntWritable, WeekBoroughWritable, AvgAccidentsAndLetalWritable> {
 	// as daily avg of incidents and lethal incidents per week per borough
 
 	public void reduce(WeekBoroughWritable key, Iterator<IntWritable> values,
-			OutputCollector<WeekBoroughWritable, AvgAccidentsAndLetalWritable> output, Reporter reporter) throws IOException {
+			OutputCollector<WeekBoroughWritable, AvgAccidentsAndLetalWritable> output, Reporter reporter)
+			throws IOException {
 		int accidents = 0;
 		int lethal_accidents = 0;
-		while(values.hasNext()){
+		while (values.hasNext()) {
 			accidents++;
-			lethal_accidents+=values.next().get();
+			lethal_accidents += values.next().get();
 		}
-		float avg_accidents = accidents / 7;
-		float avg_lethal_accidents = lethal_accidents / 7;
+		float avg_accidents = ((float) accidents) / 7;
+		float avg_lethal_accidents = ((float) lethal_accidents) / 7;
 		output.collect(key, new AvgAccidentsAndLetalWritable(avg_accidents, avg_lethal_accidents));
 	}
